pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "polyphony"
version = "0.1.0"
description = "Collaborative qualitative data analysis with human + LLM coders"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
keywords = ["qualitative research", "QDA", "LLM", "coding", "social science"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Information Analysis",
]
dependencies = [
    "click>=8.1",
    "rich>=13.7",
    "pydantic>=2.6",
    "ollama>=0.2",
    "krippendorff>=0.6",
    "scikit-learn>=1.4",
    "numpy>=1.26",
    "pandas>=2.2",
    "PyYAML>=6.0",
    "python-docx>=1.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "mypy>=1.9",
]

[project.scripts]
polyphony = "polyphony.cli.main:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["polyphony*"]

[tool.setuptools.package-data]
polyphony = ["db/migrations/*.sql"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
